# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒资管理系统 REST API。成功响应直接返回业务 JSON（camelCase）；失败返回 { err_code, err_msg }。",
        "title": "GIN-MAM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/transcode/callback": {
            "post": {
                "description": "外部转码服务 HMAC 签名回调，无需 JWT",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "转码回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HMAC 签名",
                        "name": "X-Transcode-Sign",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "回调体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.CallbackInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/group": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "创建转码组",
                "parameters": [
                    {
                        "description": "转码组参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.createGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.GroupVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/group/page": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "转码组分页",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.GroupPageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/group/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "更新转码组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "转码组 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.updateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.GroupVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "转码"
                ],
                "summary": "删除转码组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "转码组 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/task": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建任务并提交外部转码服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "创建转码任务",
                "parameters": [
                    {
                        "description": "任务参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.createTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.TaskVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/task/page": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "转码任务分页",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.TaskPageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/transcode/task/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "转码"
                ],
                "summary": "获取转码任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_transcode.TaskVO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo": {
            "type": "object",
            "properties": {
                "err_code": {
                    "type": "integer",
                    "example": 40100
                },
                "err_msg": {
                    "type": "string",
                    "example": "未授权"
                }
            }
        },
...
        "internal_domain_transcode.CallbackInput": {
            "type": "object",
            "properties": {
                "errorMsg": {
                    "type": "string"
                },
                "outputPath": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "taskId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_transcode.GroupPageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_transcode.GroupVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_transcode.GroupVO": {
            "type": "object",
            "properties": {
                "availableFlag": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "defaultFlag": {
                    "type": "integer"
                },
                "groupType": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "strategyType": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_domain_transcode.TaskPageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_transcode.TaskVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_transcode.TaskVO": {
            "type": "object",
            "properties": {
                "assetFileId": {
                    "type": "string"
                },
                "assetId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "errorMsg": {
                    "type": "string"
                },
                "externalJobId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "outputPath": {
                    "type": "string"
                },
                "profileId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "transcodeGroupId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_domain_transcode.createGroupRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "availableFlag": {
                    "type": "integer"
                },
                "defaultFlag": {
                    "type": "integer"
                },
                "groupType": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "strategyType": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_transcode.createTaskRequest": {
            "type": "object",
            "required": [
                "assetId",
                "transcodeGroupId"
            ],
            "properties": {
                "assetFileId": {
                    "type": "string"
                },
                "assetId": {
                    "type": "string"
                },
                "transcodeGroupId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_transcode.pageRequest": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "internal_domain_transcode.updateGroupRequest": {
            "type": "object",
            "properties": {
                "availableFlag": {
                    "type": "integer"
                },
                "defaultFlag": {
                    "type": "integer"
                },
                "groupType": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "strategyType": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}